requirements.txt
nltk                      3.8.1
numpy                     1.26.3
pillow                    10.2.0
wordcloud                 1.9.3
beautifulsoup4            4.12.2
pandas                    2.1.4